separate neptulon-client into its own module
