+++ .idea/modules.xml
@@ -4,6 +4,7 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/nbusy-android.iml" filepath="$PROJECT_DIR$/nbusy-android.iml" />
+      <module fileurl="file://$PROJECT_DIR$/neptulon-client/neptulon-client.iml" filepath="$PROJECT_DIR$/neptulon-client/neptulon-client.iml" />
     </modules>
   </component>
 </project>
