+++ app/build.gradle
@@ -33,11 +33,6 @@ dependencies {
     compile 'com.google.android.gms:play-services-identity:8.3.0'
     compile 'com.google.android.gms:play-services-plus:8.3.0'
     compile 'com.google.android.gms:play-services-gcm:8.3.0'
-    compile 'com.google.code.gson:gson:2.5'
     compile 'com.google.guava:guava:19.0'
-    compile 'com.squareup.okhttp3:okhttp:3.0.0-RC1'
-    compile 'com.squareup.okhttp3:okhttp-ws:3.0.0-RC1'
-    testCompile 'junit:junit:4.12'
-    testCompile 'org.mockito:mockito-core:1.10.19'
-    testCompile 'org.hamcrest:hamcrest-library:1.3'
+    compile project(':neptulon-client')
 }
