+++ neptulon-client/src/main/java/neptulon/client/Request.java
@@ -1,4 +1,4 @@
-package neptulon;
+package neptulon.client;
 
 /**
  * JSON-RPC request object.
