+++ neptulon-client/src/main/java/neptulon/client/ResHandler.java
@@ -1,4 +1,4 @@
-package neptulon;
+package neptulon.client;
 
 import com.google.gson.Gson;
