+++ neptulon-client/src/main/java/neptulon/client/Response.java
@@ -1,4 +1,4 @@
-package neptulon;
+package neptulon.client;
 
 /**
  * JSON-RPC response object.
