+++ neptulon-client/src/main/java/neptulon/client/middleware/Logger.java
@@ -1,4 +1,4 @@
-package neptulon.middleware;
+package neptulon.client.middleware;
 
 /**
  * Neptulon request/response logger middleware.
