+++ neptulon-client/src/test/java/neptulon/client/NeptulonTest.java
@@ -1,4 +1,4 @@
-package neptulon;
+package neptulon.client;
 
 import org.junit.Test;
